cleanup(core): Move `JsRealm` into a separate module
This will help make reviews easier for #15760, which moves a number of
methods related to module loading from `JsRuntime` into `JsRealm`.
